Refactor HTTP POST methods into .util.io
