+++ plugins/com.google.cloud.tools.eclipse.usagetracker.test/src/com/google/cloud/tools/eclipse/usagetracker/AnalyticsPingManagerTest.java
@@ -16,6 +16,8 @@
 
 package com.google.cloud.tools.eclipse.usagetracker;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anyBoolean;
 import static org.mockito.Matchers.eq;
@@ -26,20 +28,17 @@ import static org.mockito.Mockito.when;
 
 import com.google.cloud.tools.eclipse.preferences.AnalyticsPreferences;
 import com.google.cloud.tools.eclipse.usagetracker.AnalyticsPingManager.PingEvent;
-
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.concurrent.ConcurrentLinkedQueue;
 import org.eclipse.core.runtime.preferences.IEclipsePreferences;
 import org.eclipse.swt.widgets.Display;
-import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.mockito.verification.VerificationMode;
-
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.concurrent.ConcurrentLinkedQueue;
 
 @RunWith(MockitoJUnitRunner.class)
 public class AnalyticsPingManagerTest {
@@ -49,13 +48,15 @@ public class AnalyticsPingManagerTest {
   private static final String EVENT_TYPE = "some-event-type";
   private static final String EVENT_NAME = "some-event-name";
 
+  private static final String VIRTUAL_HOST = "virtual.host";
   private static final String VIRTUAL_DOCUMENT_PAGE =
-      "/virtual/some-application/" + EVENT_TYPE + "/" + EVENT_NAME;
+      "/virtual/" + EVENT_TYPE + "/" + EVENT_NAME;
 
   private static final String METADATA_KEY = "some-custom-key";
   private static final String METADATA_VALUE = "some-custom-value";
 
-  @SuppressWarnings("serial")
+  // TODO(chanseok): add a test to that makes an actual HTTP request (locally) and checks the
+  // incoming request. (https://github.com/GoogleCloudPlatform/google-cloud-eclipse/issues/1437)
   private static final Map<String, String> RANDOM_PARAMETERS = Collections.unmodifiableMap(
       new HashMap<String, String>() {
         {
@@ -69,27 +70,9 @@ public class AnalyticsPingManagerTest {
           put("cid", UUID);
           put("cd19", EVENT_TYPE);
           put("cd20", EVENT_NAME);
+          put("dh", VIRTUAL_HOST);
           put("dp", VIRTUAL_DOCUMENT_PAGE);
           put("dt", METADATA_KEY + "=" + METADATA_VALUE);
-        }
-      });
-
-  @SuppressWarnings("serial")
-  private static final Map<String, String> ENCODED_PARAMETERS = Collections.unmodifiableMap(
-      new HashMap<String, String>() {
-        {
-          put("dt", "some-custom-key%3Dsome-custom-value");
-          put("cd16", "0");
-          put("cd17", "0");
-          put("v", "1");
-          put("t", "pageview");
-          put("cd21", "1");
-          put("cd20", "some-event-name");
-          put("ni", "0");
-          put("tid", "UA-12345678-1");
-          put("dp", "%2Fvirtual%2Fsome-application%2Fsome-event-type%2Fsome-event-name");
-          put("cid", "bee5d838-c3f8-4940-a944-b56973597e74");
-          put("cd19", "some-event-type");
         }
       });
 
@@ -98,45 +81,24 @@ public class AnalyticsPingManagerTest {
   @Mock private ConcurrentLinkedQueue<PingEvent> pingEventQueue;
 
   @Test
-  public void testGetParametersString() {
-    String urlEncodedParameters = AnalyticsPingManager.getParametersString(RANDOM_PARAMETERS);
-
-    String[] keyValuePairs = urlEncodedParameters.split("&");
-    Assert.assertEquals(keyValuePairs.length, RANDOM_PARAMETERS.size());
-
-    for (String pair : keyValuePairs) {
-      String[] keyValue = pair.split("=");
-      Assert.assertEquals(2, keyValue.length);
-      Assert.assertEquals(keyValue[1], ENCODED_PARAMETERS.get(keyValue[0]));
-    }
-  }
-
-  @Test
-  public void testGetParametersString_percentEscaping() {
-    Map<String, String> noEscape = new HashMap<>();
-    noEscape.put("k", ".*-_abcXYZ");
-    Assert.assertEquals("k=.*-_abcXYZ", AnalyticsPingManager.getParametersString(noEscape));
-
-    Map<String, String> escape = new HashMap<>();
-    escape.put("k", " ü한글+=,`~!@#$%^&()?<>{}][|:;/\\'\"");
-    Assert.assertEquals("k=+%C3%BC%ED%95%9C%EA%B8%80%2B%3D%2C%60%7E%21%40%23"
-        + "%24%25%5E%26%28%29%3F%3C%3E%7B%7D%5D%5B%7C%3A%3B%2F%5C%27%22",
-        AnalyticsPingManager.getParametersString(escape));
-  }
-
-  @Test
   public void testEventTypeEventNameConvention() {
     PingEvent event = new PingEvent("some.event-name", null, null, null);
     Map<String, String> parameters = AnalyticsPingManager.buildParametersMap("clientId", event);
-    Assert.assertEquals("/virtual/gcloud-eclipse-tools/some.event-name",
-        parameters.get("dp"));
+    assertEquals("/virtual/gcloud-eclipse-tools/some.event-name", parameters.get("dp"));
+  }
+
+  @Test
+  public void testVirtualHostSet() {
+    PingEvent event = new PingEvent("some.event-name", null, null, null);
+    Map<String, String> parameters = AnalyticsPingManager.buildParametersMap("clientId", event);
+    assertTrue(parameters.get("dh").startsWith("virtual."));
   }
 
   @Test
   public void testMetadataConvention() {
     PingEvent event = new PingEvent("some.event-name", "times-happened", "1234", null);
     Map<String, String> parameters = AnalyticsPingManager.buildParametersMap("clientId", event);
-    Assert.assertEquals("times-happened=1234", parameters.get("dt"));
+    assertEquals("times-happened=1234", parameters.get("dt"));
   }
 
   @Test
