+++ plugins/com.google.cloud.tools.eclipse.usagetracker/META-INF/MANIFEST.MF
@@ -9,9 +9,8 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Export-Package: com.google.cloud.tools.eclipse.usagetracker
 Require-Bundle: org.eclipse.core.runtime;bundle-version="3.11.1",
  org.eclipse.jface,
- org.eclipse.ui.workbench,
- com.google.cloud.tools.eclipse.preferences,
- com.google.cloud.tools.eclipse.util
-Import-Package: com.google.common.annotations;version="[20.0.0,21.0.0)",
- com.google.common.escape;version="[20.0.0,21.0.0)",
- com.google.common.net;version="[20.0.0,21.0.0)"
+ org.eclipse.ui.workbench
+Import-Package: com.google.cloud.tools.eclipse.preferences,
+ com.google.cloud.tools.eclipse.util,
+ com.google.cloud.tools.eclipse.util.io,
+ com.google.common.annotations;version="[20.0.0,21.0.0)"
