+++ plugins/com.google.cloud.tools.eclipse.util/META-INF/MANIFEST.MF
@@ -11,8 +11,11 @@ Export-Package: com.google.cloud.tools.eclipse.util,
  com.google.cloud.tools.eclipse.util.service,
  com.google.cloud.tools.eclipse.util.status,
  com.google.cloud.tools.eclipse.util.templates.appengine
-Import-Package: com.google.common.base;version="[20.0.0,21.0.0)",
+Import-Package: com.google.common.annotations;version="[20.0.0,21.0.0)",
+ com.google.common.base;version="[20.0.0,21.0.0)",
+ com.google.common.escape;version="[20.0.0,21.0.0)",
  com.google.common.io;version="[20.0.0,21.0.0)",
+ com.google.common.net;version="[20.0.0,21.0.0)",
  freemarker.template;version="[2.3.25,2.4.0)",
  org.eclipse.core.commands,
  org.eclipse.core.resources,
